+++ CHANGES.rst
@@ -8,6 +8,7 @@ Development
 - [ui] Make IssueReporter work on dev and prod equally good
 - [ui] Improve Stack subsystem mechanics
 - [ui] Improve logging and debugging
+- [ui] Add ApplicationProfile subsystem
 
 
 2018-11-23 0.160.0